+++ CHANGES.rst
@@ -5,7 +5,12 @@ Changelog
 1.0.0a2 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Switch checkbox widget to use z3c.form.browser.checkbox to be compatible
+  with Plone 5.0.x
+  [csenger]
+
+- Add Plone4 support for banner and registration form
+  [csenger]
 
 
 1.0.0a1 (2018-05-21)